feat: add fee channel and remove outdated notes
- Removed outdated notes regarding fee explanation and database transaction flow schema.
- Deleted unnecessary Excalidraw files related to donation flow and wallet flow.
- Cleaned up mental model notes for multiple currencies to clarify storage and display principles.

+++ internal/infra/postgres/migrations/000011_create_table_wallet_ledger_entries.up.sql
@@ -30,7 +30,7 @@ CREATE TABLE wallet_ledger_entries (
 
   -- copied snapshot for audit
   fee_fixed BIGINT NOT NULL DEFAULT 0,
-  fee_percentage NUMERIC(5,4) NOT NULL DEFAULT 0,
+  fee_percentage BIGINT NOT NULL DEFAULT 0,
   fee_amount BIGINT NOT NULL DEFAULT 0,
   net_amount BIGINT NOT NULL,
